Add some minor changes to classes
