+++ TowerForge/TowerForge/Commons/Enums/StatisticName.swift
@@ -1,0 +1,15 @@
+//
+//  StatisticEnums.swift
+//  TowerForge
+//
+//  Created by Rubesh on 11/4/24.
+//
+
+import Foundation
+
+enum StatisticName: String, Codable, CaseIterable {
+    case defaultStatistic
+    case totalKills
+    case totalGamesPlayed
+    case totalDeaths
+}
